dcbot/requirements.txt
discord.py==2.3.2
requests==2.31.0
flask==2.3.3
aiohttp==3.8.5